Add files via upload
We consider a system of individuals (the nature of which is irrelevant) that can take two states A and B. At any point in time, an individual can only be in one state. There are no other states possible. We will denote by [A] the number of individuals in state A and by [B] the number of individuals in state B. There are N individuals in total. All individuals can be assumed to be in contact with all other individuals. At time t = 0, there are B0 individuals in state B. All other individuals are therefore in state A. Individuals in state A turn into state B at a rate β [B] N , i.e., proportional to the number of their neighbours that are in state B. Individuals in state B turn into state A at a fixed rate γ.